[Kath Pino] [#2.4 ] Actualization Menu CLI and add UserService Class, User class and more

+++ src/Menu.java
@@ -29,6 +29,8 @@ public class Menu {
     public void RecibeOptionChooseForUserAndCallTheActionSelect(String option) {
         Scanner scanner = new Scanner(System.in);
         boolean verification= false;
+        String credential;
+        String password;
         switch (option){
             case "0":
                 System.out.println("Quit Application");
@@ -38,31 +40,47 @@ public class Menu {
                 PrintCollectionBooks();
                 break;
             case "2":
-                System.out.println("Checkout Book \n");
-                System.out.println("List of Books:  \n");
-                PrintCollectionBooks();
-                System.out.println("----------------------------------------------");
-                System.out.println("Name of Book: \n");
-                String nameBook= scanner.nextLine();
-                verification = BookingService.RemoveBookToTheSystem(nameBook);
-                PrintMessageAboutCheckout(verification);
-                PrintCollectionBooks();
+                System.out.println("Checkout Book \n" + "User: \n");
+                credential= scanner.nextLine();
+                System.out.println("Password: \n");
+                password= scanner.nextLine();
+                if(UserService.VerificationOfUserAdminAndPass(credential, password)) {
+                    System.out.println("List of Books:  \n");
+                    PrintCollectionBooks();
+                    System.out.println("----------------------------------------------");
+                    System.out.println("Name of Book: \n");
+                    String nameBook= scanner.nextLine();
+                    verification = BookingService.RemoveBookToTheSystem(nameBook);
+                    PrintMessageAboutCheckout(verification);
+                    PrintCollectionBooks();
+                }
+                else{
+                    System.out.println("Credential fail");
+                }
                 break;
             case "3":
-                System.out.println("Return Book \n");
-                System.out.println("List of Books:  \n");
-                PrintCollectionBooks();
-                System.out.println("----------------------------------------------");
-                System.out.println("Name of Book: \n");
-                String nameOfBook= scanner.nextLine();
-                System.out.println("Date: \n");
-                String date= scanner.nextLine();
-                System.out.println("Author: \n");
-                String author= scanner.nextLine();
-                Book book = new Book(nameOfBook, date,author);
-                verification = BookingService.AddBookToTheSystem(book);
-                PrintMessageAboutAddBook(verification);
-                PrintCollectionBooks();
+                System.out.println("Return Book \n" + "User: \n");
+                credential= scanner.nextLine();
+                System.out.println("Password: \n");
+                password= scanner.nextLine();
+                if(UserService.VerificationOfUserAdminAndPass(credential, password)) {
+                    System.out.println("List of Books:  \n");
+                    PrintCollectionBooks();
+                    System.out.println("----------------------------------------------");
+                    System.out.println("Name of Book: \n");
+                    String nameOfBook = scanner.nextLine();
+                    System.out.println("Date: \n");
+                    String date = scanner.nextLine();
+                    System.out.println("Author: \n");
+                    String author = scanner.nextLine();
+                    Book book = new Book(nameOfBook, date, author);
+                    verification = BookingService.AddBookToTheSystem(book);
+                    PrintMessageAboutAddBook(verification);
+                    PrintCollectionBooks();
+                }
+                else{
+                    System.out.println("Credential Fail");
+                }
                 break;
             case "4":
                 System.out.println("List of Movies \n");
@@ -118,6 +136,8 @@ public class Menu {
             return false;
         }
     }
+
+
 
 
 }
